Moved image calibration scripts to ScanAnalysis
This is a better home
