+++ ScanAnalysis/calibrations/Undulator/calibration_scripts/magspec_charge_calibration/single_scan_charge_calibration_acavemagcam3.py
@@ -15,10 +15,10 @@ import numpy as np
 import matplotlib.pyplot as plt
 from scipy.optimize import curve_fit
 from pathlib import Path
-from ImageAnalysis.image_analysis.utils import read_imaq_png_image
-import ImageAnalysis.image_analysis.analyzers.calibration_scripts.modules_image_processing.shot_charge_reader as charge_reader
-from ImageAnalysis.image_analysis.analyzers.UC_GenericMagSpecCam import UC_GenericMagSpecCamAnalyzer
-import ImageAnalysis.image_analysis.analyzers.online_analysis_modules.directory_functions as directory_functions
+import sys
+from image_analysis.utils import read_imaq_png_image
+import calibrations.Undulator.shot_charge_reader as charge_reader
+from image_analysis.analyzers.UC_GenericMagSpecCam import UC_GenericMagSpecCamAnalyzer
 
 
 def linear(x, a):
@@ -33,40 +33,20 @@ if sampleCase == 1:
     data_day = 9
     data_month = 5
     data_year = 2024
-    scan_number = 51  # 30, 50, 51
-    image_name = "UC_HiResMagCam"
-
-elif sampleCase == 101:
-    data_day = 29
-    data_month = 6
-    data_year = 2023
-    scan_number = 23
-    image_name = "U_HiResMagCam"
-
-elif sampleCase == 102:
-    data_day = 25
-    data_month = 7
-    data_year = 2023
-    scan_number = 24
-    image_name = "U_HiResMagCam"
-
+    scan_number = 54  # 30, 50, 51
+    image_name = "UC_ACaveMagCam3"
 else:
-    print("Pick a valid sample case!")
-    data_day = 0
-    data_month = 0
-    data_year = 0
-    scan_number = 0
-    image_name = None
+    sys.exit()
 
 super_path = directory_functions.compile_daily_path(data_day, data_month, data_year)
 
 calibration_analyzer = UC_GenericMagSpecCamAnalyzer(
-    mag_spec_name='hires',
-    noise_threshold=100,
-    roi=[1, -1, 1, -1],
+    mag_spec_name='acave3',
+    noise_threshold=4,
+    roi=[112, 278, 23, 1072],
     saturation_value=4095,
-    normalization_factor=1,  # 6.218486e-6  OLD: 7.643283839778091e-07,
-    transverse_calibration=43,
+    normalization_factor=1.0,
+    transverse_calibration=129.4,
     do_transverse_calculation=False,  # True,
     transverse_slice_threshold=0.02,
     transverse_slice_binsize=5,
@@ -82,7 +62,7 @@ clipping_arr = np.zeros(num_shots)
 saturation_arr = np.zeros(num_shots)
 camera_counts_arr = np.zeros(num_shots)
 
-ict_name = 'U_BCaveICT'
+ict_name = 'U_UndulatorExitICT'
 tdms_filepath = charge_reader.compile_tdms_filepath(super_path, scan_number)
 picoscope_charge_arr = charge_reader.get_beam_charge(tdms_filepath, device=ict_name, channel=ict_name + ' Charge pC')
 
@@ -108,6 +88,9 @@ for i in range(len(shot_arr)):
         print("Saturation Counts:", saturation_arr[i])
         print("Camera Counts:", camera_counts_arr[i])
         print("Picoscope Charge:", picoscope_charge_arr[i])
+
+        plt.imshow(image)
+        plt.show()
 print()
 
 clip_tolerance = 0.001
@@ -129,7 +112,7 @@ print()
 if normalizationCheck:
     min_camera = 2
 else:
-    min_camera = 2e6
+    min_camera = 1e6
 
 both_pass = np.where((clipping_arr < clip_tolerance) &
                      (saturation_arr < sat_tolerance) &
@@ -146,7 +129,7 @@ print("Fit Values:", linear_fit)
 
 print("Calibration Constants:")
 tdms_filepath = charge_reader.compile_tdms_filepath(super_path, scan_number)
-charge_reader.print_normalization(num_shots, tdms_filepath)
+# charge_reader.print_normalization(num_shots, tdms_filepath, image_name=image_name)
 print("const_normalization_factor =", linear_fit[0])
 
 axis = np.linspace(min(camera_counts_arr), max(camera_counts_arr), 50)
